src/com/pennant/prodmtr/test/AnalyticServiceImplTest.java
package com.pennant.prodmtr.test;

import static org.mockito.ArgumentMatchers.anyInt;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.when;

import java.util.ArrayList;
import java.util.Date;
import java.util.List;

import org.testng.Assert;
import org.testng.annotations.BeforeClass;
import org.testng.annotations.Test;

import com.pennant.prodmtr.Dao.Interface.AnalyticsDao;
import com.pennant.prodmtr.model.Dto.AnalyticsDto;
import com.pennant.prodmtr.model.Dto.ProjectDto;
import com.pennant.prodmtr.model.Entity.Module;
import com.pennant.prodmtr.model.Entity.Project;
import com.pennant.prodmtr.model.Entity.ProjectTask;
import com.pennant.prodmtr.model.Entity.Subtask;
import com.pennant.prodmtr.model.Entity.Task;
import com.pennant.prodmtr.model.view.ModuleSummary;
import com.pennant.prodmtr.model.view.ProjectSummary;
import com.pennant.prodmtr.model.view.SubtaskSummary;
import com.pennant.prodmtr.model.view.TaskSummary;
import com.pennant.prodmtr.model.view.TotalAnalysis;
import com.pennant.prodmtr.service.Impl.AnalyticServiceImpl;

public class AnalyticServiceImplTest {

    private AnalyticsDao analyticsDao;
    private AnalyticServiceImpl analyticService;

    @BeforeClass
    public void setUp() {
        // Create a mock AnalyticsDao
        analyticsDao = mock(AnalyticsDao.class);

        // Initialize AnalyticServiceImpl with the mock AnalyticsDao
        analyticService = new AnalyticServiceImpl(analyticsDao);
    }

    @Test
    public void testGetUserProjDetails() {
        // Prepare mock data
        int userId = 1;
        List<ProjectTask> projectTasks = new ArrayList<>();
        List<Task> tasks = new ArrayList<>();
        List<Subtask> subtasks = new ArrayList<>();
        // Add mock data to lists...

        // Mock behavior of the AnalyticsDao methods
        when(analyticsDao.getAllProjectsWorkingHoursByUid(anyInt())).thenReturn(projectTasks);
        when(analyticsDao.getAllProjectsIndvTasksWorkingHoursByUid(anyInt())).thenReturn(tasks);
        when(analyticsDao.getAllSubTasksByUid(anyInt())).thenReturn(subtasks);
        // Add more mock behaviors...

        // Call the method to be tested
        AnalyticsDto result = analyticService.getUserProjDetails(userId);

        // Perform assertions on the result
        // For example:
        Assert.assertNotNull(result);
        // Add more assertions based on the expected result from the mock data...
    }

    @Test
    public void testGetCompletedTasksByDate() {
        // Prepare mock data
        List<Task> tasks = new ArrayList<>();
        List<ProjectTask> projectTasks = new ArrayList<>();
        // Add mock data to lists...

        // Mock behavior of the AnalyticsDao methods
        when(analyticsDao.getAllCompletedTasks()).thenReturn(tasks);
        when(analyticsDao.getAllCompletedProjectTasks()).thenReturn(projectTasks);
        // Add more mock behaviors...

        // Call the method to be tested
        List<String> result = analyticService.getCompletedTasksByDate();

        // Perform assertions on the result
        // For example:
        Assert.assertNotNull(result);
        // Add more assertions based on the expected result from the mock data...
    }

    @Test
    public void testGetTotalAnalysis() {
        // Prepare mock data
        List<ProjectDto> projects = new ArrayList<>();
        List<Module> modules = new ArrayList<>();
        List<Task> subtasks = new ArrayList<>();
        List<Task> tasks = new ArrayList<>();
        List<ProjectTask> projectTasks = new ArrayList<>();
        // Add mock data to lists...

        // Mock behavior of the AnalyticsDao methods
        when(analyticsDao.getAllProjects()).thenReturn(projects);
        when(analyticsDao.getAllModules()).thenReturn(modules);
        when(analyticsDao.getAllCompletedTasks()).thenReturn(subtasks);
        when(analyticsDao.getAllTasks()).thenReturn(tasks);
        when(analyticsDao.getAllProjTasks()).thenReturn(projectTasks);
        // Add more mock behaviors...

        // Call the method to be tested
        TotalAnalysis result = analyticService.getTotalAnalysis();

        // Perform assertions on the result
        // For example:
        Assert.assertNotNull(result);
        // Add more assertions based on the expected result from the mock data...
    }

    // Write more test methods for other public methods of the AnalyticServiceImpl class...
}
